Add files via upload
This will ask one question from the data randomly. It waits for you to understand and imagine the answer, you can write the answer as well . Then you press the answer and it gives you correct answer. There is no scoring, just a practice before you play the practice game.